+++ TablePro/Models/Query/QueryTab.swift
@@ -50,7 +50,7 @@ struct QueryTab: Identifiable, Equatable {
         self.pendingChanges = TabChangeSnapshot()
         self.selectedRowIndices = []
         self.sortState = SortState()
-        self.filterState = TabFilterState()
+        self.filterState = TabFilterState(isVisible: tabType == .table)
         self.columnLayout = ColumnLayoutState()
         self.pagination = PaginationState()
         self.hasUserInteraction = false
@@ -82,7 +82,7 @@ struct QueryTab: Identifiable, Equatable {
         self.pendingChanges = TabChangeSnapshot()
         self.selectedRowIndices = []
         self.sortState = SortState()
-        self.filterState = TabFilterState()
+        self.filterState = TabFilterState(isVisible: persisted.tabType == .table)
         self.columnLayout = ColumnLayoutState()
         self.pagination = PaginationState()
         self.hasUserInteraction = false